Move pixel sending logic inside AtbInitializer
